version_1/MLOPS/End To End ML Project/15-ETL/15-1-ETLPipeline-Project-1/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-http
apache-airflow-providers-postgres


